Fixing autocomplete
