+++ lookbook/app/controllers/auto_complete_test_controller.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+
+class AutoCompleteTestController < ApplicationController
+  def index
+    @fruit_list = [
+      'Apples',
+      'Apricots',
+      'Avocado',
+      'Ackee',
+      'Bananas',
+      'Bilberries',
+      'Blueberries',
+      'Blackberries',
+      'Boysenberries',
+      'Bread fruit',
+      'Cantaloupes (cantalope)',
+      'Chocolate-Fruit',
+      'Cherimoya',
+      'Cherries',
+      'Cranberries',
+      'Cucumbers',
+      'Currants',
+      'Dates',
+      'Durian',
+      'Eggplant',
+      'Elderberries',
+      'Figs',
+      'Gooseberries',
+      'Grapes',
+      'Grapefruit',
+      'Guava'
+    ].select { |fruit| fruit.include?(params['q']) }
+  end
+end
