+++ lookbook/app/controllers/preview_controller.rb
@@ -1,3 +1,4 @@
 class PreviewController < ViewComponentsController
   helper Lookbook::PreviewHelper
+  helper Primer::ViewHelper
 end
