+++ lookbook/app/views/auto_complete_test/index.html.erb
@@ -1,0 +1,3 @@
+<% @fruit_list.each do |fruit| %>
+  <%= render(Primer::Beta::AutoComplete::Item.new(value: fruit)) { fruit } %>
+<% end %>
